+++ app/javascript/dashboard/i18n/locale/en/inboxMgmt.json
@@ -97,8 +97,8 @@
         "SUBMIT_BUTTON": "Create inbox"
       },
       "TWILIO": {
-        "TITLE": "Twilio SMS/Whatsapp Channel",
-        "DESC": "Integrate Twilio and start supporting your customers via SMS or Whatsapp.",
+        "TITLE": "Twilio SMS/WhatsApp Channel",
+        "DESC": "Integrate Twilio and start supporting your customers via SMS or WhatsApp.",
         "ACCOUNT_SID": {
           "LABEL": "Account SID",
           "PLACEHOLDER": "Please enter your Twilio Account SID",
@@ -114,8 +114,8 @@
           "ERROR": "This field is required"
         },
         "CHANNEL_NAME": {
-          "LABEL": "Channel Name",
-          "PLACEHOLDER": "Please enter a channel name",
+          "LABEL": "Inbox Name",
+          "PLACEHOLDER": "Please enter a inbox name",
           "ERROR": "This field is required"
         },
         "PHONE_NUMBER": {
@@ -136,13 +136,14 @@
         "TITLE": "SMS Channel via Twilio",
         "DESC": "Start supporting your customers via SMS with Twilio integration."
       },
-       "WHATSAPP_VIA_TWILIO": {
-        "TITLE": "Whatsapp Channel via Twilio",
-        "DESC": "Start supporting your customers via Whatsapp with Twilio integration."
-       },
 			"WHATSAPP": {
-        "TITLE": "Whatsapp Channel",
-        "DESC": "Start supporting your customers via Whatsapp.",
+        "TITLE": "WhatsApp Channel",
+        "DESC": "Start supporting your customers via WhatsApp.",
+        "PROVIDERS": {
+          "LABEL": "API Provider",
+          "TWILIO": "Twilio",
+          "360_DIALOG": "360Dialog"
+        },
         "INBOX_NAME": {
           "LABEL": "Inbox Name",
           "PLACEHOLDER": "Please enter an inbox name",
@@ -155,12 +156,15 @@
         },
         "API_KEY": {
           "LABEL": "API key",
-          "SUBTITLE": "Configure the whatsapp API key.",
-          "PLACEHOLDER": "API key"
+          "SUBTITLE": "Configure the WhatsApp API key.",
+          "PLACEHOLDER": "API key",
+          "APPLY_FOR_ACCESS": "Don't have any API key? Apply for access here",
+          "ERROR": "Please enter a valid value."
+
         },
-        "SUBMIT_BUTTON": "Create Whatsapp Channel",
+        "SUBMIT_BUTTON": "Create WhatsApp Channel",
         "API": {
-          "ERROR_MESSAGE": "We were not able to save the whatsapp channel"
+          "ERROR_MESSAGE": "We were not able to save the WhatsApp channel"
         }
       },
       "API_CHANNEL": {
@@ -244,7 +248,7 @@
       },
       "AUTH": {
         "TITLE": "Choose a channel",
-        "DESC": "Chatwoot supports live-chat widget, Facebook page, Twitter profile, Whatsapp, Email etc., as channels. If you want to build a custom channel, you can create it using the API channel. Select one channel from the options below to proceed."
+        "DESC": "Chatwoot supports live-chat widget, Facebook page, Twitter profile, WhatsApp, Email etc., as channels. If you want to build a custom channel, you can create it using the API channel. Select one channel from the options below to proceed."
       },
       "AGENTS": {
         "TITLE": "Agents",
